+++ ipaserver/install/server/upgrade.py
@@ -22,7 +22,9 @@ from augeas import Augeas
 
 from ipalib import api, x509
 from ipalib.constants import RENEWAL_CA_NAME, RA_AGENT_PROFILE, IPA_CA_RECORD
-from ipalib.install import certmonger, sysrestore
+from ipalib.install import certmonger
+from ipalib import sysrestore
+from ipalib.facts import is_ipa_configured
 import SSSDConfig
 import ipalib.util
 import ipalib.errors
@@ -1484,7 +1486,7 @@ def upgrade_configuration():
     fstore = sysrestore.FileStore(paths.SYSRESTORE)
     sstore = sysrestore.StateFile(paths.SYSRESTORE)
 
-    if installutils.is_ipa_configured() is None:
+    if is_ipa_configured() is None:
         sstore.backup_state('installation', 'complete', True)
 
     fqdn = api.env.host
